Procedural generation of maps is quite solid now
Right now the main thing to be solved is that despite attempting to center the map every time a room is created (thus maximizing development space), the map seems to only get so large before having problems. It's likely something to do with the how the currentColumn and currentRow are updated after centering the map.